Extract FsDriver transactions into it's own module.
